Revert "Move some diagnostic code to the type checker"
This reverts commit 0e6d43c552060e02bb1b00a5708e1496b06f4e92.
