+++ src/providers/diagnostics/typeInferenceDiagnostics.ts
@@ -5,16 +5,26 @@ import {
   CodeActionKind,
   CodeActionParams,
   Diagnostic,
+  DiagnosticSeverity,
+  Range,
   TextEdit,
 } from "vscode-languageserver";
+import { SyntaxNode } from "web-tree-sitter";
+import { PositionUtil } from "../../positionUtil";
 import { TreeUtils } from "../../util/treeUtils";
 import { URI } from "vscode-uri";
 import { ElmWorkspaceMatcher } from "../../util/elmWorkspaceMatcher";
+import { mapSyntaxNodeToExpression } from "../../util/types/expressionTree";
 import { MultistepOperation } from "../../util/multistepOperation";
+import { TypeChecker } from "../../util/types/typeChecker";
 import { ITreeContainer } from "../../forest";
 import { IElmWorkspace } from "../../elmWorkspace";
-import { ThrottledCancellationToken } from "../../cancellation";
+import {
+  ICancellationToken,
+  ThrottledCancellationToken,
+} from "../../cancellation";
 import { container } from "tsyringe";
+import { Diagnostics, error } from "../../util/types/diagnostics";
 
 export class TypeInferenceDiagnostics {
   TYPE_INFERENCE = "Type Inference";
@@ -67,7 +77,8 @@ export class TypeInferenceDiagnostics {
             }
 
             diagnostics.push(
-              ...checker.getDiagnosticsFromDeclaration(
+              ...this.getDiagnosticsForDeclaration(
+                checker,
                 allTopLevelFunctions[index],
                 treeContainer,
                 new ThrottledCancellationToken(cancellationToken),
@@ -90,5 +101,52 @@ export class TypeInferenceDiagnostics {
         reject,
       );
     });
+  }
+
+  public getDiagnosticsForDeclaration(
+    checker: TypeChecker,
+    declaration: SyntaxNode,
+    treeContainer: ITreeContainer,
+    cancellationToken: ICancellationToken,
+  ): Diagnostic[] {
+    const valueDeclaration = mapSyntaxNodeToExpression(declaration);
+    if (valueDeclaration?.nodeType !== "ValueDeclaration") {
+      return [];
+    }
+
+    const diagnostics: Diagnostic[] = [];
+
+    checker
+      .getDiagnosticsFromDeclaration(valueDeclaration, cancellationToken)
+      .forEach((diagnostic) => {
+        const nodeUri = (<any>diagnostic.data).uri;
+
+        if (nodeUri === treeContainer.uri) {
+          diagnostics.push(diagnostic);
+        }
+      });
+
+    if (!valueDeclaration.typeAnnotation) {
+      const typeString: string = checker.typeToString(
+        checker.findType(declaration),
+        treeContainer,
+      );
+
+      if (
+        typeString &&
+        typeString !== "unknown" &&
+        declaration.firstNamedChild?.firstNamedChild
+      ) {
+        diagnostics.push(
+          error(
+            declaration.firstNamedChild.firstNamedChild,
+            Diagnostics.MissingTypeAnnotation,
+            typeString,
+          ),
+        );
+      }
+    }
+
+    return diagnostics;
   }
 }
